complete libft -V1.00

+++ libft/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: ayoub <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/20 18:14:58 by ayoub             #+#    #+#             */
-/*   Updated: 2024/08/20 18:14:59 by ayoub            ###   ########.fr       */
+/*   Updated: 2024/08/20 22:44:05 by ayoub            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned int	i;
 	unsigned char	*p;
